float price

+++ backend/zayna/models.py
@@ -82,9 +82,9 @@ class Project(models.Model):
     MODE_CHOICES = Modes
 
     name = models.CharField(default="", null=False, blank=True, max_length=255)
-    price = models.IntegerField(default=0)
+    price = models.FloatField(default=0)
     price_by_level = models.JSONField(default=list)
-    income = models.IntegerField(default=0)
+    income = models.FloatField(default=0)
     income_by_level = models.JSONField(default=list)
     payment = models.IntegerField(default=0)
     mode = models.CharField(max_length=12, choices=MODE_CHOICES.choices, default=Modes.MARKET, null=True)
